requirements.txt
alembic==0.9.1
appdirs==1.4.3
beautifulsoup4==4.5.3
bs4==0.0.1
chardet==3.0.2
click==6.7
cssselect==1.0.1
dominate==2.3.1
Flask==0.12.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
html2text==2016.9.19
itsdangerous==0.24
Jinja2==2.9.6
lxml==3.7.3
Mako==1.0.6
MarkupSafe==1.0
olefile==0.44
packaging==16.8
Pillow==4.1.0
pyparsing==2.2.0
python-editor==1.0.3
readability-lxml==0.6.2
requests==2.13.0
simplejson==3.10.0
six==1.10.0
SQLAlchemy==1.1.9
visitor==0.1.3
Werkzeug==0.12.1
WTForms==2.1
